chore: add configuration of fileupload
