+++ src/main/java/com/khureturn/community/service/AccountService.java
@@ -125,4 +125,21 @@ public class Oauth2LoginService {
             return new ResponseEntity<MemberResponseDto.GoogleAccountInfoDto>(googleAccountInfoDto, httpHeaders, HttpStatus.OK);
         }
     }
+
+
+    public void logout(String accessToken, HttpServletResponse httpServletResponse) throws IOException {
+
+        String getAccessToken = jwtTokenProvider.extractAccessToken(accessToken).orElse(null);
+
+        String username = jwtTokenProvider.getUsername(getAccessToken);
+
+        Cookie cookie= new Cookie("authorization", null);
+        cookie.setMaxAge(0);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        httpServletResponse.addCookie(cookie);
+
+        Long expiration = jwtTokenProvider.getExpireTime(getAccessToken);
+        redisService.setBlackListToken(getAccessToken, "BLACKLIST_ACCESSTOKEN_" + username, expiration);
+    }
 }
